feat(navbar): add menu
