+++ web/src/common/components/navbar/Navbar.tsx
@@ -1,43 +1,12 @@
-import { supabase } from '@/supabase'
-import { LogoutOutlined } from '@ant-design/icons'
-import { User } from '@supabase/supabase-js'
-import { Avatar, Button, Flex, Typography } from 'antd'
-import gravatar from 'gravatar'
 import React from 'react'
+import { NavbarUser } from './NavbarUser'
+import { NavigationOptions } from './NavigationOptions'
 
 export const Navbar: React.FC = () => {
-  const [user, setUser] = React.useState<User | null>(null)
-  React.useEffect(() => {
-    const getUser = async () => {
-      const { data } = await supabase.auth.getUser()
-      if (data) {
-        setUser(data.user)
-      }
-    }
-    getUser()
-  }, [])
-
-  const userName = user?.email ?? 'keine EMail'
-  const greeting = user?.email === 'griebner@mediainterface.de' ? 'Bonjour' : 'Hello'
-  const greetingText = `${greeting} ${userName}`
-
-  const gravatarUrl = gravatar.url(user?.email ?? '')
-
-  const handleLogOut = async () => {
-    await supabase.auth.signOut()
-  }
-
   return (
     <>
-      <Flex justify={'flex-end'} align={'flex-start'}>
-        <Typography style={{ color: 'hotpink', marginTop: '5px', marginRight: '15px' }}>{greetingText}</Typography>
-
-        <Avatar src={gravatarUrl} />
-
-        <Button type="text" icon={<LogoutOutlined />} onClick={handleLogOut} style={{ color: 'hotpink' }}>
-          Logout
-        </Button>
-      </Flex>
+      <NavigationOptions />
+      <NavbarUser />
     </>
   )
 }
