+++ web/src/features/auth/presentation/AuthProvider.tsx
@@ -3,6 +3,7 @@ import { Auth } from '@supabase/auth-ui-react'
 import { ThemeSupa } from '@supabase/auth-ui-shared'
 
 
+import { Navbar } from '@/common/components/navbar/Navbar'
 import { ROUTING_PATH } from '@/features/router/domain/constants/routing-path.constants'
 import React from 'react'
 import { Outlet, useNavigate } from 'react-router-dom'
@@ -39,7 +40,10 @@ export const AuthProvider: React.FC = () => {
   return shouldShowAuthScreen ? (
     <Auth supabaseClient={supabase} providers={[]} redirectTo="" appearance={{ theme: ThemeSupa }} />
   ) : (
-    <Outlet />
+    <>
+      <Navbar />
+      <Outlet />
+    </>
   )
 }
